Split JniValueManager and ReflectionJniValueManager
